Changed RaceHorse float data types to double.

+++ src/main/java/com/djuwidja/horsegame/pathfinder/meta/RaceHorse.java
@@ -3,15 +3,16 @@ package com.djuwidja.horsegame.pathfinder.meta;
 import lombok.Getter;
 
 public class RaceHorse {
-	@Getter private float fwdSpdMax; // max forward speed
-	@Getter private float angSpdMax; // max angular speed
-	@Getter private float fwdAcc; // forward acceleration
-	@Getter private float angAcc; // angular acceleration
+	@Getter private double fwdSpdMax; // max forward speed
+	@Getter private double angSpdMax; // max angular speed
+	@Getter private double fwdAcc; // forward acceleration
+	@Getter private double angAcc; // angular acceleration
 	
-	public RaceHorse(float fwdSpdMax,
-					 float angSpdMax,
-					 float fwdAcc,
-					 float angAcc) {
+	public RaceHorse(
+			double fwdSpdMax,
+			double angSpdMax,
+			double fwdAcc,
+			double angAcc) {
 		this.fwdSpdMax = fwdSpdMax;
 		this.angSpdMax = angSpdMax;
 		this.fwdAcc = fwdAcc;
